requirements.txt
streamlit
google-generativeai
matplotlib
textblob